refactor: streamline exports query validation and improve integer coercion
